pyproject.toml
[project]
name = "hass-energy"
version = "0.1.0"
description = "Energy management system integrating Home Assistant data with MILP planning"
readme = "README.md"
authors = [
    { name = "Nick Whyte", email = "nick@nickwhyte.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "httpx>=0.27.2",
    "pyyaml>=6.0.2",
    "pulp>=2.7.0",
    "matplotlib>=3.9.2",
]

[project.scripts]
hass-energy = "hass_energy.cli:main"

[dependency-groups]
dev = [
    "pyright>=1.1.390",
    "ruff>=0.7.3",
]

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["hass_energy"]
